Rename Interface to Interfaces in namespaces
