+++ src/core/Microsoft.PowerToys.Settings.UI.Lib/Interfaces/ISettingsRepository`1.cs
@@ -2,7 +2,7 @@
 // The Microsoft Corporation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Microsoft.PowerToys.Settings.UI.Lib.Interface
+namespace Microsoft.PowerToys.Settings.UI.Lib.Interfaces
 {
     public interface ISettingsRepository<T>
     {
