+++ src/core/Microsoft.PowerToys.Settings.UI.Lib/PowerRenameSettings.cs
@@ -4,7 +4,7 @@
 
 using System;
 using System.Text.Json.Serialization;
-using Microsoft.PowerToys.Settings.UI.Lib.Interface;
+using Microsoft.PowerToys.Settings.UI.Lib.Interfaces;
 
 namespace Microsoft.PowerToys.Settings.UI.Lib
 {
